+++ NEWS.md
@@ -12,7 +12,8 @@
   - `version_lt()`: Check whether a schema version property is less than.
 * Move `as_config()` from `hubAdmin` to `hubUtils` package (#173).
 * `<config>` class objects now have a `type` attribute to track what type of config they contain (i.e `"tasks"` or `"admin"`).
-* Attempt to convert output of `read_config()` and `read_config_file()` to a `<config>` class object (#173).
+* Attempt to convert output of `read_config()` and `read_config_file()` to a `<config>` class object (#173).`
+* Fix bug in `extract_schema_version()` where only single digits from each version component were being extracted.
   
 
 # hubUtils 0.1.7