+++ NEWS.md
@@ -1,4 +1,4 @@
-# hubUtils (development version)
+# hubUtils 0.2.0
 
 * Add family of functions for extracting the version number from a variety of sources:
   - `get_version_config()`: extract version from a `<config>` class object.